Retrieve messages from DB adapter

+++ src/main/java/com/chg/hackdays/chappie/server/service/SessionService.java
@@ -1,7 +1,7 @@
 package com.chg.hackdays.chappie.server.service;
 
 public interface SessionService {
-    int getCurrentConversationId();
+    Long getCurrentConversationId();
 
     String getUsername();
 }
